Some fixes
* Changed the regexes for barcoding analysis. There was a main issue
  related to the presence of both R1/2 tag and pipeline tags (i.e.:
  .trimmed). The new approach should work flawlessly even with more
  complex cases;
* input_filenames is not evaluated two times when using
  last_operation_filenames;
* Splitting by sample id should not be performed anymore (it is an
  obsolete behaviour)
* Better exception handling for variables evaluation inside Executor
* sort_tempdir could not exist in case we are using --skip-mapping
  parameter

+++ hatspil/mapping.py
@@ -268,7 +268,8 @@ class Mapping:
             unlink_inputs=True
         )
 
-        shutil.rmtree(self.sort_tempdir)
+        if os.path.exists(self.sort_tempdir):
+            shutil.rmtree(self.sort_tempdir)
         self.analysis.logger.info("Finished sorting")
 
     def mark_duplicates(self):
